# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Indexer APIs for Andamio dapp-system.",
        "title": "Andamio Indexer 1.0.0",
        "contact": {
            "name": "Andamio Support",
            "url": "https://www.andamio.com/support",
            "email": "dev@andamio.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "0.0.0.0:42069",
    "basePath": "/api/v1/indexer",
    "paths": {
        "/addresses": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds a new address to the indexer for monitoring transactions and UTxOs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Add Address",
                "operationId": "addAddress",
                "parameters": [
                    {
                        "description": "The address object containing the address string to be added.",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodel.AddressRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully added address.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing address.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes an address from the tracking list.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Remove Address",
                "operationId": "removeAddress",
                "parameters": [
                    {
                        "description": "The address object containing the address string to be removed.",
                        "name": "addressRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/viewmodel.AddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully removed address.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing address.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/addresses/{address}/assets": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of assets held at a specific address, with support for pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Get Assets by Address",
                "operationId": "getAssetsByAddress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The address to retrieve assets for.",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved assets.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.Asset"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid address or pagination parameters.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found or no assets found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/addresses/{address}/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves transactions associated with a specific address with pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Get Transactions by Address",
                "operationId": "getTransactionsByAddress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The address to retrieve transactions for.",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transactions.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid address or pagination parameters.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found or no transactions found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "viewmodel.AddressRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
        "viewmodel.Asset": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "fingerprint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_hex": {
                    "type": "string"
                },
                "policy_id": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Datum": {
            "type": "object",
            "properties": {
                "datum_cbor": {
                    "type": "string"
                },
                "datum_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Redeemer": {
            "type": "object",
            "properties": {
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "tag": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "type": "string"
                }
            }
        },
        "viewmodel.SimpleUTxO": {
            "type": "object",
            "properties": {
                "transaction_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Transaction": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "block_number": {
                    "type": "integer"
                },
                "certificates": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fee": {
                    "type": "integer"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.TransactionInput"
                    }
                },
                "metadata": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.TransactionOutput"
                    }
                },
                "reference_inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.SimpleUTxO"
                    }
                },
                "slot_number": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "ttl": {
                    "type": "integer"
                },
                "withdrawals": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "witness": {
                    "$ref": "#/definitions/viewmodel.Witness"
                }
            }
        },
        "viewmodel.TransactionInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "asset": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Asset"
                    }
                },
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "datum": {
                    "$ref": "#/definitions/viewmodel.Datum"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.TransactionOutput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "asset": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Asset"
                    }
                },
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "datum": {
                    "$ref": "#/definitions/viewmodel.Datum"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
...
        "viewmodel.Witness": {
            "type": "object",
            "properties": {
                "plutus_data": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v1_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v2_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v3_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "redeemers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Redeemer"
                    }
                },
                "transaction_hash": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}